// lib/prisma/enums.ts
export const ROLE = {
  SYSTEM_OWNER: "SYSTEM_OWNER",
  SYSTEM_ADMIN: "SYSTEM_ADMIN",
  SYSTEM_USER: "SYSTEM_USER",
  USER: "USER",
} as const;

export type ROLE = (typeof ROLE)[keyof typeof ROLE];

export const SESSION_DEVICE_TYPE = {
  DESKTOP: "DESKTOP",
  TABLET: "TABLET",
  MOBILE: "MOBILE",
  CONSOLE: "CONSOLE",
  EMBEDDED: "EMBEDDED",
  SMARTTV: "SMARTTV",
  WEARABLE: "WEARABLE",
  XR: "XR",
  OTHER: "OTHER",
} as const;

export type SESSION_DEVICE_TYPE =
  (typeof SESSION_DEVICE_TYPE)[keyof typeof SESSION_DEVICE_TYPE];

export const USER_SYSTEM_ACCESS_LEVEL = {
  VIEW: "VIEW",
  APPROVE: "APPROVE",
  SUSPEND: "SUSPEND",
  CREATE: "CREATE",
  UPDATE: "UPDATE",
  DELETE: "DELETE",
} as const;

export type USER_SYSTEM_ACCESS_LEVEL =
  (typeof USER_SYSTEM_ACCESS_LEVEL)[keyof typeof USER_SYSTEM_ACCESS_LEVEL];